pyproject.toml
[tool.poetry]
name = "lms-homework"
version = "0.1.0"
description = ""
authors = ["lacry <lacryk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.9"
drf-yasg = "^1.21.8"
stripe = "^11.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
